# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZMABT API",
    "version": "1.0.0",
    "description": "API documentation for ZMABT backend"
  },
  "servers": [
    {
      "url": "http://localhost:3333/api/v1/",
      "description": "Local API server"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 64
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "password": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 32
                  },
                  "passwordConfirmation": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 32
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password",
                  "passwordConfirmation"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 64
                  },
                  "password": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 32
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns access token"
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout user",
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "User not authenticated"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/account/profile": {
      "get": {
        "summary": "Get user profile",
        "description": "Without userId returns the authenticated user's own profile. With userId returns another user's profile and whether the authenticated user follows them.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" },
            "description": "ID of the user to view. Omit to get own profile."
          }
        ],
        "responses": {
          "200": {
            "description": "Own profile (no userId)",
            "content": {
              "application/json": {
                "examples": {
                  "ownProfile": {
                    "summary": "Own profile (no userId)",
                    "value": {
                      "user": {
                        "id": 1,
                        "username": "johndoe",
                        "email": "john@example.com",
                        "bio": "...",
                        "createdAt": "2026-01-01T00:00:00.000Z",
                        "updatedAt": "2026-01-01T00:00:00.000Z"
                      }
                    }
                  },
                  "otherProfile": {
                    "summary": "Other user's profile (with userId)",
                    "value": {
                      "user": {
                        "id": 5,
                        "username": "janedoe",
                        "email": "jane@example.com",
                        "bio": "...",
                        "profilePicturePath": "uploads/abc.jpg",
                        "createdAt": "2026-01-01T00:00:00.000Z",
                        "updatedAt": "2026-01-01T00:00:00.000Z"
                      },
                      "isFollowing": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/create": {
      "put": {
        "summary": "Create a new post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "postText": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "latitude": {
                    "type": "number"
                  }
                },
                "required": [
                  "postText",
                  "rating",
                  "longitude",
                  "latitude"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/getPageFyp": {
      "get": {
        "summary": "Get FYP posts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated posts"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/getPage": {
      "get": {
        "summary": "Get posts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated posts"
          },
          "422": {
            "description": "Invalid user id"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/getPageUser": {
      "get": {
        "summary": "Get posts of authenticated user",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user's posts"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/getPagePlace": {
      "get": {
        "summary": "Get paginated posts for a place",
        "parameters": [
          {
            "name": "placeId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated posts for a place"
          },
          "422": {
            "description": "Invalid place id"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/get": {
      "get": {
        "summary": "Get a single post by ID",
        "parameters": [
          {
            "name": "postId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the post"
          },
          "422": {
            "description": "Post id invalid"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/comments/create": {
      "post": {
        "summary": "Create a new comment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "postId": {
                    "type": "integer"
                  },
                  "commentId": {
                    "type": "integer",
                    "nullable": true
                  }
                },
                "required": [
                  "content",
                  "postId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment created successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "422": {
            "description": "Parent comment does not belong to the specified post"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/comments/getPage": {
      "get": {
        "summary": "Get paginated comments for a post or parent comment",
        "parameters": [
          {
            "name": "postId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "commentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "required": false
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated comments"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Failed to load comments"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/comments/update": {
      "patch": {
        "summary": "Update a comment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "commentId": {
                    "type": "integer"
                  },
                  "content": {
                    "type": "string",
                    "maxLength": 5000
                  }
                },
                "required": [
                  "commentId",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "403": {
            "description": "Cannot edit comments of other users"
          },
          "404": {
            "description": "Comment not found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/update": {
      "patch": {
        "summary": "Update a post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "postId": {
                    "type": "integer"
                  },
                  "postText": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  }
                },
                "required": [
                  "postId",
                  "postText",
                  "rating"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "403": {
            "description": "Cannot edit posts of other users"
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/delete": {
      "delete": {
        "summary": "Delete a post",
        "parameters": [
          {
            "name": "postId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "403": {
            "description": "Cannot delete posts of other users"
          },
          "422": {
            "description": "Post id does is invalid"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/account/follow": {
      "post": {
        "summary": "Follow a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer"
                  }
                },
                "required": ["userId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User followed successfully"
          },
          "400": {
            "description": "Cannot follow yourself"
          },
          "401": {
            "description": "User not authenticated"
          },
          "422": {
            "description": "User id is Invalid"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/account/unfollow": {
      "post": {
        "summary": "Unfollow a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer"
                  }
                },
                "required": ["userId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User unfollowed successfully"
          },
          "400": {
            "description": "Cannot unfollow yourself"
          },
          "401": {
            "description": "User not authenticated"
          },
          "422": {
            "description": "User id is invalid"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/comments/delete": {
      "delete": {
        "summary": "Delete a comment",
        "parameters": [
          {
            "name": "commentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "403": {
            "description": "Cannot delete comments of other users"
          },
          "404": {
            "description": "Comment not found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/comments/like": {
      "put": {
        "summary": "Like a comment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "commentId": {
                    "type": "integer"
                  }
                },
                "required": ["commentId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment liked successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "422": {
            "description": "Comment not found"
          }
        }
      }
    },
    "/comments/removeLike": {
      "delete": {
        "summary": "Remove like from a comment",
        "parameters": [
          {
            "name": "commentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment like removed successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "422": {
            "description": "Comment not found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ratings/set": {
      "post": {
        "summary": "Create or update a rating for a post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "postId": {
                    "type": "integer"
                  },
                  "stars": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  }
                },
                "required": ["postId", "stars"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rating created or updated successfully"
          },
          "400": {
            "description": "Invalid rating data"
          },
          "401": {
            "description": "User not authenticated"
          },
          "403": {
            "description": "You cannot rate your own post"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/places": {
      "get": {
        "summary": "Get all places",
        "description": "Returns all places with their coordinates. Used to render pins on the map.",
        "responses": {
          "200": {
            "description": "List of all places",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "latitude": { "type": "number" },
                      "longitude": { "type": "number" },
                      "aiDescription": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/places/info": {
      "get": {
        "summary": "Get place detail with average rating",
        "description": "Returns a single place by ID or coordinates, along with its average star rating.",
        "parameters": [
          {
            "name": "placeId",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "name": "latitude",
            "in": "query",
            "required": false,
            "schema": { "type": "number" }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": false,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the place and its average rating",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "place": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "id": { "type": "integer" },
                        "latitude": { "type": "number" },
                        "longitude": { "type": "number" },
                        "aiDescription": { "type": "string" },
                        "createdAt": { "type": "string", "format": "date-time" }
                      }
                    },
                    "rating": { "type": "number", "nullable": true }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
    "/account/updateBio": {
      "patch": {
        "summary": "Update authenticated user's bio",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bio": {
                    "type": "string",
                    "maxLength": 250
                  }
                },
                "required": ["bio"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bio updated successfully"
          },
          "401": {
            "description": "User not authenticated"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}